refactor(rest): extract admin settings/cli-token routes into routes/admin_settings.py
Moves the four admin settings + CLI-token routes (GET/POST
settings/{key}, POST auth/cli-token, POST auth/cli-token/revoke) out of
create_app's closure, the first slice of the admin router split.

Lifts publish_capability_requirements and the CAPABILITY_REQUIREMENTS_SETTING
constant onto routes/deps.py; create_app keeps the old
_publish_capability_requirements name bound for the lifespan-startup
publish still in the closure.
